Add basic class for facade tests #17
Run with python -m unittest app/model/test/facade.py

+++ app/model/database.py
@@ -6,8 +6,11 @@ db = SQLAlchemy()
 
 def configure_database(app, config):
     """Set up the database with the given flask app."""
-
-    app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///' + config['database-path']
+    if len(config['database-path']) > 0:
+        app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///' + config['database-path']
+    else:
+        # in memory
+        app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite://'
     app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 
     # setup sqlalchemy for flask
@@ -15,14 +18,14 @@ def configure_database(app, config):
 
     # delete database on debug launch
     if config['debug'] and config['debug-db-reset']:
-        if os.path.exists(config['database-path']):
+        if len(config['database-path']) > 0 and os.path.exists(config['database-path']):
             os.remove(config['database-path'])
         # drop everything
         with app.app_context():
             db.drop_all()
 
     # create database if it does not exist
-    if not os.path.exists(config['database-path']):
+    if len(config['database-path']) == 0 or not os.path.exists(config['database-path']):
         with app.app_context():
             db.create_all()
